trying to make components adaptive for different  screens size
