+++ app/src/main/java/com/gdgevents/gdgeventsapp/MainActivity.kt
@@ -26,18 +26,3 @@ class MainActivity : ComponentActivity() {
 }
 
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    GDGEventsAppTheme {
-        Greeting("Android")
-    }
-}
