+++ app/src/main/java/com/gdgevents/gdgeventsapp/features/onBoarding/components/OnBoardItem.kt
@@ -3,19 +3,37 @@ package com.gdgevents.gdgeventsapp.features.onBoarding.components
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gdgevents.gdgeventsapp.features.onBoarding.model.OnBoardModel
-import com.gdgevents.gdgeventsapp.features.onBoarding.model.onBoardModel
 import com.gdgevents.gdgeventsapp.ui.theme.GDGEventsAppTheme
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.Dp
+import com.gdgevents.gdgeventsapp.features.onBoarding.model.onBoardList
 
 @Composable
 fun OnBoardItem(
@@ -24,38 +42,179 @@ fun OnBoardItem(
     onSkipClick: () -> Unit = {},
     onNextClick: () -> Unit = {},
 ) {
-    Box(
-        modifier = modifier.fillMaxSize()
-            .padding(top = 20.dp),
+
+
+    BoxWithConstraints(
+        modifier = modifier
+            .fillMaxHeight()
     ) {
+        val screenHeight = maxHeight
+        val screenWidth = maxWidth
+        val dynamicPadding = screenHeight * 0.15f // 15% of the screen height for padding
+        val dynamicImageSize = screenWidth * 0.6f // 50% of screen width for image size
+        val dynamicTextPadding = screenHeight * 0.07f
+
         Image(
             painter = painterResource(id = page.imageRes),
             contentDescription = null,
-            modifier = modifier
-                .height(570.dp)
-                .width(360.dp)
-                .padding(top = 80.dp)
+            modifier = Modifier
+                .padding(
+                    start = 16.dp,
+                    end = 16.dp,
+                    top = dynamicPadding
+                )
+                .fillMaxWidth(0.9f)
+                .aspectRatio(0.7f)
+                .size(dynamicImageSize)
                 .align(Alignment.TopCenter),
         )
-        Column(modifier = modifier.align(Alignment.BottomCenter)) {
+        Column(
+            modifier = Modifier.align(Alignment.BottomCenter)
+        ) {
             OnboardingTextContainer(
                 title = page.title,
                 description = page.description,
                 currentPage = page.id,
                 totalPages = 2,
-                onNextClick = {onNextClick()},
-                onSkipClick = {onSkipClick()},
+                dyanmicPaddind = dynamicTextPadding,
+                onNextClick = { onNextClick() },
+                onSkipClick = { onSkipClick() },
             )
         }
+    }
+}
+
+@Composable
+fun OnboardingTextContainer(
+    title: String,
+    description: String,
+    currentPage: Int,
+    totalPages: Int,
+    dyanmicPaddind: Dp,
+    onNextClick: () -> Unit,
+    onSkipClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Box(
+        modifier = modifier
+            .fillMaxWidth()
+            .background(
+                color = MaterialTheme.colorScheme.primary,
+                shape = RoundedCornerShape(topStart = 24.dp, topEnd = 24.dp)
+            )
+            .padding(top = dyanmicPaddind, bottom = dyanmicPaddind, start = 20.dp, end = 20.dp),
+        contentAlignment = Alignment.BottomEnd
+
+    ) {
+        Column {
+            Text(
+                text = title,
+                textAlign = TextAlign.Center,
+                style = MaterialTheme.typography.headlineLarge,
+                color = MaterialTheme.colorScheme.onPrimary,
+                modifier = modifier
+                    .align(Alignment.CenterHorizontally),
+            )
+            Text(
+                text = description,
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onPrimary,
+                modifier = modifier.padding(top = 8.dp, bottom = 20.dp),
+                textAlign = TextAlign.Center,
+                maxLines = 2
+            )
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(9.dp)
+            ) {
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    modifier = modifier
+                        .weight(2f)
+                        .align(Alignment.CenterVertically)
+                        .padding(start = 16.dp),
+                    horizontalArrangement = Arrangement.Center
+                ) {
+                    repeat(totalPages) { index ->
+                        val color = if (index == currentPage)
+                            MaterialTheme.colorScheme.onPrimary
+                        else MaterialTheme.colorScheme.secondary
+                        Spacer(modifier = modifier.width(4.dp))
+                        Box(
+                            modifier = modifier
+                                .height(8.dp)
+                                .width(if (index == currentPage) 33.dp else 8.dp)
+                                .background(color, CircleShape)
+                                .padding(horizontal = 4.dp),
+                        )
+                    }
+                }
+                Text(
+                    text = "Next",
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.onPrimary,
+                    textAlign = TextAlign.End,
+                    modifier = modifier
+                        .clickable { onNextClick() }
+                )
+            }
+        }
+    }
+}
+
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_6
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_3A
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_2
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_3A
+)
+@Composable
+private fun OnBoardModelPrev() {
+    GDGEventsAppTheme {
+        OnBoardItem(onBoardList[0])
     }
 
 }
 
-@Preview(showBackground = true, showSystemUi = true)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_6
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_3A
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_2
+)
+@Preview(
+    showBackground = true,
+    showSystemUi = true,
+    device = Devices.PIXEL_3A
+)
 @Composable
-private fun OnBoardModelPrev() {
+private fun OnBoardModelPrev1() {
     GDGEventsAppTheme {
-        OnBoardItem(onBoardModel[0])
+        OnBoardItem(onBoardList[1])
     }
 
 }
